+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/NPC/ChaosNPCLifecycle.lua
@@ -291,7 +291,7 @@ function ChaosNPC:OnZombieDead()
     local isFollowGroup = self.npcGroup == ChaosNPCGroupID.COMPANIONS or
         self.npcGroup == ChaosNPCGroupID.FOLLOWERS
     if isFollowGroup and self.zombie then
-        ChaosUtils.PlayUISound("deathcrash", false, 0.5)
+        ChaosUtils.PlayUISound("deathcrash", false, 0.25)
 
         local player = getPlayer()
         if player then